Authentication is Compleated

+++ routers/todos.py
@@ -4,6 +4,7 @@ from sqlalchemy.orm import Session
 from typing import Annotated
 from database import  SessionLocal
 import models
+from .auth  import get_current_user
 from models import Todos
 
 router = APIRouter()
@@ -17,6 +18,7 @@ def get_db():
         db.close()
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency=Annotated[dict,Depends(get_current_user)]
 
 # This class is Used for Validation 
 class TodoRequest(BaseModel):
@@ -27,22 +29,26 @@ class TodoRequest(BaseModel):
 
 # Read all the books in Database
 @router.get("/", status_code=status.HTTP_200_OK)
-async def read_all(db: db_dependency):
+async def read_all(user:user_dependency,db: db_dependency):
     # Ensure query returns list, not ORM objects directly
-    return db.query(models.Todos).all()
+    return db.query(models.Todos).filter(Todos.owner_id==user.get('id')).all()
 
 
 @router.get("/todos/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).first()
+async def read_todo(user:user_dependency,db: db_dependency, todo_id: int = Path(gt=0)):
+    todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).filter(models.Todos.owner_id==user.get('id')).first()
     if todo_model is not None:
         return todo_model
-    raise HTTPException(status_code=404, detail="Id not found.")
+    raise HTTPException(status_code=404, detail="Todos not found.")
 
 
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo_request: TodoRequest):
-    todo_model = models.Todos(**todo_request.dict())
+async def create_todo(user:user_dependency, db: db_dependency, 
+                      todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=401,detail='Authentication Failed')
+    
+    todo_model = models.Todos(**todo_request.dict(),owner_id=user.get('id'))
     db.add(todo_model)
     db.commit()
     db.refresh(todo_model)   # refresh to get new ID
@@ -50,26 +56,33 @@ async def create_todo(db: db_dependency, todo_request: TodoRequest):
 
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_data(db: db_dependency,
+async def update_todos(user: user_dependency,
+                      db: db_dependency,
                       todo_request: TodoRequest,
                       todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+    
+    todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).filter(models.Todos.owner_id == user.get('id')).first()
 
-    todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo is Not Found")
-    else:
-        todo_model.title = todo_request.title
-        todo_model.description = todo_request.description
-        todo_model.priority = todo_request.priority
-        todo_model.complete = todo_request.complete
-        db.add(todo_model)
-        db.commit()
-        db.refresh(todo_model)   # refresh after update
+    
+    todo_model.title = todo_request.title
+    todo_model.description = todo_request.description
+    todo_model.priority = todo_request.priority
+    todo_model.complete = todo_request.complete
+    db.add(todo_model)
+    db.commit()
 
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).first()
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+
+    todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).filter(models.Todos.owner_id == user.get('id')).first()
+
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found")
     db.delete(todo_model)
